Add TwoPaneLayout component
