+++ src/sections/planner/Planner.tsx
@@ -1,8 +1,7 @@
-import { useEffect, useMemo, useRef, useState } from 'react';
 
 import { getNextDay } from '@common/utils/date_utils.ts';
-import Storage from '@common/utils/storage';
 import { StorageKeys } from '@common/utils/storage_keys.ts';
+import TwoPaneLayout from '@common/components/TwoPaneLayout/TwoPaneLayout.tsx';
 import PlannerAgendas from '@planner/components/PlannerAgendas/PlannerAgendas.tsx';
 import PlannerDay from '@planner/components/PlannerDay/PlannerDay.tsx';
 import DateSelector from '@planner/components/DateSelector/DateSelector.tsx';
@@ -13,33 +12,9 @@ import { useSelectedDate } from '@planner/hooks/useSelectedDate.ts';
 const DEFAULT_LEFT_PANE_WIDTH_PERCENT = 66.6667; // w-2/3
 const MIN_LEFT_PANE_WIDTH_PERCENT = 30;
 const MAX_LEFT_PANE_WIDTH_PERCENT = 70;
-const RESIZE_HANDLE_WIDTH_PX = 6;
-
-const mdQuery = '(min-width: 768px)';
 
 const Planner = () => {
   const { selectedDate, setSelectedDate } = useSelectedDate();
-
-  // Track if viewport is at least md to apply resizable width only on md+
-  const [isMd, setIsMd] = useState<boolean>(() => {
-    if (typeof window === 'undefined') {
-      // assume md on SSR to avoid layout shift
-      return true;
-    }
-    return window.matchMedia(mdQuery).matches;
-  });
-  useEffect(() => {
-    if (typeof window === 'undefined') {
-      return;
-    }
-
-    const mql = window.matchMedia(mdQuery);
-    const onChange = () => setIsMd(mql.matches);
-    onChange();
-    mql.addEventListener?.('change', onChange);
-
-    return () => mql.removeEventListener?.('change', onChange);
-  }, []);
 
   const {
     plannerAgendas,
@@ -71,127 +46,47 @@ const Planner = () => {
     handleSnoozeDayItem,
   } = usePlannerDays(selectedDate);
 
-  // Resizable split state
-  const plannerContainerRef = useRef<HTMLDivElement | null>(null);
-  const [leftPaneWidthPercent, setLeftPaneWidthPercent] = useState<number>(() => {
-    if (typeof window === 'undefined') {
-      return DEFAULT_LEFT_PANE_WIDTH_PERCENT;
-    }
-
-    const savedWidthRaw = Storage.get(StorageKeys.PlannerLeftPaneWidth, null);
-    const savedWidthParsed = savedWidthRaw ? parseFloat(savedWidthRaw) : NaN;
-    if (isNaN(savedWidthParsed)) {
-      return DEFAULT_LEFT_PANE_WIDTH_PERCENT;
-    }
-
-    return Math.min(
-      MAX_LEFT_PANE_WIDTH_PERCENT,
-      Math.max(MIN_LEFT_PANE_WIDTH_PERCENT, savedWidthParsed),
-    );
-  });
-  const [isDragging, setIsDragging] = useState(false);
-
-  const handleMouseMove = (clientX: number) => {
-    const container = plannerContainerRef.current;
-    if (!container) {
-      return;
-    }
-
-    const rect = container.getBoundingClientRect();
-    const available = rect.width - RESIZE_HANDLE_WIDTH_PX;
-    const xWithin = Math.min(Math.max(clientX - rect.left, 0), available);
-    const percent = (xWithin / available) * 100;
-    const clamped = Math.min(
-      MAX_LEFT_PANE_WIDTH_PERCENT,
-      Math.max(MIN_LEFT_PANE_WIDTH_PERCENT, percent),
-    );
-    setLeftPaneWidthPercent(clamped);
-  };
-
-  useEffect(() => {
-    if (typeof window === 'undefined' || !isDragging) {
-      return;
-    }
-
-    const onMouseMove = (e: MouseEvent) => handleMouseMove(e.clientX);
-    const onMouseUp = () => {
-      setIsDragging(false);
-      Storage.set(StorageKeys.PlannerLeftPaneWidth, String(leftPaneWidthPercent));
-    };
-
-    window.addEventListener('mousemove', onMouseMove);
-    window.addEventListener('mouseup', onMouseUp);
-    window.addEventListener('mouseleave', onMouseUp);
-
-    return () => {
-      window.removeEventListener('mousemove', onMouseMove);
-      window.removeEventListener('mouseup', onMouseUp);
-      window.removeEventListener('mouseleave', onMouseUp);
-    };
-  }, [isDragging, leftPaneWidthPercent]);
-
-  const leftPaneStyle = useMemo(
-    () => ({ width: `${leftPaneWidthPercent}%` }),
-    [leftPaneWidthPercent],
-  );
-
   return (
-    <div
-      ref={plannerContainerRef}
-      className="md:h-screen flex flex-col md:flex-row w-full bg-bg-main overflow-auto md:overflow-hidden"
-    >
-      {/* Left pane (days) */}
-      <div
-        className="w-full md:flex-none md:overflow-auto md:min-h-0 flex flex-col"
-        style={isMd ? leftPaneStyle : undefined}
-      >
-        <DateSelector selectedDate={selectedDate} onDateChange={setSelectedDate} />
+    <TwoPaneLayout
+      storageKey={StorageKeys.PlannerLeftPaneWidth}
+      defaultLeftWidthPercent={DEFAULT_LEFT_PANE_WIDTH_PERCENT}
+      minLeftWidthPercent={MIN_LEFT_PANE_WIDTH_PERCENT}
+      maxLeftWidthPercent={MAX_LEFT_PANE_WIDTH_PERCENT}
+      leftPane={
+        <>
+          <DateSelector selectedDate={selectedDate} onDateChange={setSelectedDate} />
 
-        <PlannerDay
-          date={selectedDate}
-          dayItems={daysItems}
-          dragDayItem={dragDayItem}
-          onDragStartDayItem={handleDayItemDragStart}
-          onDragEndDayItem={handleDayItemDragEnd}
-          getItemsForDate={getItemsForDate}
-          onReorderDayItems={handleReorderDayItems}
-          onAddDayItem={handleAddDayItem}
-          onUpdateDayItem={handleUpdateDayItem}
-          onDeleteDayItem={handleDeleteDayItem}
-          onCopyDayItem={handleCopyDayItem}
-          onSnoozeDayItem={handleSnoozeDayItem}
-        />
-        <PlannerDay
-          date={getNextDay(selectedDate)}
-          dayItems={daysItems}
-          dragDayItem={dragDayItem}
-          onDragStartDayItem={handleDayItemDragStart}
-          onDragEndDayItem={handleDayItemDragEnd}
-          getItemsForDate={getItemsForDate}
-          onReorderDayItems={handleReorderDayItems}
-          onAddDayItem={handleAddDayItem}
-          onUpdateDayItem={handleUpdateDayItem}
-          onDeleteDayItem={handleDeleteDayItem}
-          onCopyDayItem={handleCopyDayItem}
-          onSnoozeDayItem={handleSnoozeDayItem}
-        />
-      </div>
-
-      {/* Separator with resize handle */}
-      <div
-        role="separator"
-        aria-orientation="vertical"
-        aria-label="Resize planner panes"
-        className="hidden md:block md:shrink-0 md:cursor-col-resize md:bg-transparent"
-        style={{ width: RESIZE_HANDLE_WIDTH_PX, touchAction: 'none' }}
-        onMouseDown={() => setIsDragging(true)}
-        onTouchStart={() => setIsDragging(true)}
-      >
-        <div className="h-full w-px mx-auto bg-gray-600" />
-      </div>
-
-      {/* Right pane (agendas) */}
-      <div className="w-full md:flex-1 overflow-auto min-h-0 px-8 py-5">
+          <PlannerDay
+            date={selectedDate}
+            dayItems={daysItems}
+            dragDayItem={dragDayItem}
+            onDragStartDayItem={handleDayItemDragStart}
+            onDragEndDayItem={handleDayItemDragEnd}
+            getItemsForDate={getItemsForDate}
+            onReorderDayItems={handleReorderDayItems}
+            onAddDayItem={handleAddDayItem}
+            onUpdateDayItem={handleUpdateDayItem}
+            onDeleteDayItem={handleDeleteDayItem}
+            onCopyDayItem={handleCopyDayItem}
+            onSnoozeDayItem={handleSnoozeDayItem}
+          />
+          <PlannerDay
+            date={getNextDay(selectedDate)}
+            dayItems={daysItems}
+            dragDayItem={dragDayItem}
+            onDragStartDayItem={handleDayItemDragStart}
+            onDragEndDayItem={handleDayItemDragEnd}
+            getItemsForDate={getItemsForDate}
+            onReorderDayItems={handleReorderDayItems}
+            onAddDayItem={handleAddDayItem}
+            onUpdateDayItem={handleUpdateDayItem}
+            onDeleteDayItem={handleDeleteDayItem}
+            onCopyDayItem={handleCopyDayItem}
+            onSnoozeDayItem={handleSnoozeDayItem}
+          />
+        </>
+      }
+      rightPane={
         <PlannerAgendas
           plannerAgendas={plannerAgendas}
           plannerAgendaItems={plannerAgendaItems}
@@ -209,8 +104,8 @@ const Planner = () => {
           copyAgendasMap={copyAgendasMap}
           fetchAgendaItems={fetchAgendaItems}
         />
-      </div>
-    </div>
+      }
+    />
   );
 };
